requirements.txt
openai
Pillow
python-dotenv
requests